update : modularized the resolve entity structure
